updated css
non functional, app/chatbot still not working, maybe I installed wrong...
