+++ README.txt
@@ -1,0 +1,4 @@
+install dependencies - 
+
+pip install openai
+pip install python-dotenv
