+++ templates/index.html
@@ -3,11 +3,13 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="../static/style.css">
     <title>Chatbot</title>
 </head>
 <body>
-    <h1>Chatbot</h1>
-    <div id="chat">
+    
+    <div id="chat" class="container">
+        <h1>Chatbot</h1>
         <div id="chat-box">
             {% if user_input %}
                 <p>You: {{ user_input }}</p>
@@ -17,9 +19,13 @@
             {% endif %}
         </div>
         <form id="user-input-form" action="/" method="post">
-            <input type="text" name="user_input" id="user-input" placeholder="Type your message...">
-            <button type="submit">Send</button>
+            <textarea id="inputText" placeholder="Type your input here..."></textarea>
+            <button onclick="processInput()">Submit</button>
+            <div id="output"></div>
         </form>
     </div>
+
+    <script src="../static/app.js"></script>
+
 </body>
 </html>
